+++ cspackages-test/assembly-to-lispy-test-target-nesting-container-nested-level2-nested-level3.lisp
@@ -1,17 +1,17 @@
 ;;; Generated automatically. Do not edit.
 ;;; Class: AssemblyToLispyTestTarget.NestingContainer+NestedLevel2+NestedLevel3
-;;; Generator Version: 41
-;;; Creation Date: 2026-07-11T03:36:09Z
+;;; Generator Version: 42
+;;; Creation Date: 2026-07-11T12:55:15Z
 
 (cl:in-package :assembly-to-lispy-test-target-nesting-container-nested-level2-nested-level3)
 
-(cl:defconstant <type> (dotnet:resolve-type "AssemblyToLispyTestTarget.NestingContainer+NestedLevel2+NestedLevel3"))
+(cl:define-symbol-macro <type> (dotnet:resolve-type "AssemblyToLispyTestTarget.NestingContainer+NestedLevel2+NestedLevel3"))
 (cl:defconstant <type-str> "AssemblyToLispyTestTarget.NestingContainer+NestedLevel2+NestedLevel3")
-(cl:defconstant <creation> "2026-07-11T03:36:09Z")
-(cl:defconstant <version> 41)
+(cl:defconstant <creation> "2026-07-11T12:55:15Z")
+(cl:defconstant <version> 42)
 
 ;; Register C# Type with CLOS
-(cl:eval-when (:compile-toplevel :load-toplevel :execute)
+(cl:eval-when (:load-toplevel :execute)
   (dotnet:static "DotCL.Runtime" "EnsureDotNetTypeClass"
                  (dotnet:resolve-type "AssemblyToLispyTestTarget.NestingContainer+NestedLevel2+NestedLevel3")))