+++ cspackages-test/csharp-assembly-utils.lisp
@@ -1,6 +1,6 @@
 ;;; Generated automatically. Do not edit.
-;;; Generator Version: 41
-;;; Creation Date: 2026-07-11T03:36:09Z
+;;; Generator Version: 42
+;;; Creation Date: 2026-07-11T12:55:15Z
 
 ;;; Douglas P. Fields, Jr. - symbolics@lisp.engineer
 ;;; Copyright 2026 Douglas P. Fields, Jr.