+++ cspackages-test/system-net-service-point-manager.lisp
@@ -1,17 +1,17 @@
 ;;; Generated automatically. Do not edit.
 ;;; Class: System.Net.ServicePointManager
-;;; Generator Version: 41
-;;; Creation Date: 2026-07-11T03:36:09Z
+;;; Generator Version: 42
+;;; Creation Date: 2026-07-11T12:55:15Z
 
 (cl:in-package :system-net-service-point-manager)
 
-(cl:defconstant <type> (dotnet:resolve-type "System.Net.ServicePointManager"))
+(cl:define-symbol-macro <type> (dotnet:resolve-type "System.Net.ServicePointManager"))
 (cl:defconstant <type-str> "System.Net.ServicePointManager")
-(cl:defconstant <creation> "2026-07-11T03:36:09Z")
-(cl:defconstant <version> 41)
+(cl:defconstant <creation> "2026-07-11T12:55:15Z")
+(cl:defconstant <version> 42)
 
 ;; Register C# Type with CLOS
-(cl:eval-when (:compile-toplevel :load-toplevel :execute)
+(cl:eval-when (:load-toplevel :execute)
   (dotnet:static "DotCL.Runtime" "EnsureDotNetTypeClass"
                  (dotnet:resolve-type "System.Net.ServicePointManager")))